COMP: Avoid circular import dependency
